@@ -5,14 +5,14 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>io.micronaut</groupId>
+        <groupId>io.micronaut.platform</groupId>
         <artifactId>micronaut-parent</artifactId>
-        <version>3.10.4</version>
+        <version>4.4.2</version>
     </parent>
 
     <groupId>pl.piomin.services</groupId>
     <artifactId>sample-micronaut-kubernetes</artifactId>
-    <version>1.0</version>
+    <version>1.1</version>
 
     <modules>
         <module>employee-service</module>
@@ -22,30 +22,19 @@
     <packaging>pom</packaging>
 
     <properties>
-        <jdk.version>17</jdk.version>
+        <jdk.version>21</jdk.version>
         <sonar.projectKey>piomin_sample-micronaut-kubernetes</sonar.projectKey>
         <sonar.organization>piomin</sonar.organization>
         <sonar.host.url>https://sonarcloud.io</sonar.host.url>
-        <micronaut.version>3.10.4</micronaut.version>
+        <micronaut.version>4.4.8</micronaut.version>
         <jib.maven-plugin-version>1.8.0</jib.maven-plugin-version>
+        <micronaut.test.resources.enabled>true</micronaut.test.resources.enabled>
     </properties>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>io.micronaut</groupId>
-                <artifactId>micronaut-bom</artifactId>
-                <version>${micronaut.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
 
     <build>
         <plugins>
             <plugin>
-                <groupId>io.micronaut.build</groupId>
+                <groupId>io.micronaut.maven</groupId>
                 <artifactId>micronaut-maven-plugin</artifactId>
             </plugin>
             <plugin>
